GW.Sqlbot/requirements.txt
Flask==2.3.3
pymssql==2.2.8
python-dotenv==1.0.0
openai==1.51.2